src/test/java/org/sonar/plugins/flex/flexpmd/FlexPmdRuleRepositoryTest.java
/*
 * Sonar Flex Plugin
 * Copyright (C) 2010 SonarSource
 * dev@sonar.codehaus.org
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public
 * License along with this program; if not, write to the Free Software
 * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
 */

package org.sonar.plugins.flex.flexpmd;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertThat;

import org.junit.Test;
import org.sonar.api.rules.Rule;

import java.util.List;

public class FlexPmdRuleRepositoryTest {

  @Test
  public void shouldCreateRules() {
    FlexPmdRuleRepository repository = new FlexPmdRuleRepository();
    List<Rule> rules = repository.createRules();
    assertThat(rules.size(), is(94));
    for (Rule rule : rules) {
      assertNotNull(rule.getKey());
      assertNotNull(rule.getDescription());
      assertNotNull(rule.getName());
      assertNotNull(rule.getRepositoryKey());
      assertNotNull(rule.getSeverity());
    }
  }

}
